+++ plugins/plugins.php
@@ -3,6 +3,6 @@
     * This file imports all plugins.
     * Just uncomment plugins which you want to activate
     */
-    // include_once "images/index.php";
+     include_once "images/index.php";
     include_once "text files/index.php";
 ?>